+++ src/components/contact-us/ContactUsCallToAction.tsx
@@ -1,0 +1,19 @@
+
+function ContactUsCallToAction() {
+  return (
+    <div className="flex flex-col gap-4 justify-end w-full max-w-[500px] self-center mt-6">
+        <h1 className="text-center mb-6 text-4xl font-bold">
+          Your Tech Community Awaits!
+        </h1>
+        <p className="text-center mb-4">
+          Connect with like-minded individuals, gain access to exclusive
+          resources, and grow your tech career.
+        </p>
+        <button className="w-full bg-blue-600 hover:bg-blue-700 p-3 rounded-md">
+          Join The Community
+        </button>
+      </div>
+  )
+}
+
+export default ContactUsCallToAction
